feat(website): redesign hero with aurora bloom and parallax
