+++ website/src/landing.rs
@@ -6,6 +6,7 @@ mod platform;
 mod scenes;
 
 use dioxus::prelude::*;
+use hero::Hero;
 
 pub const ICON: Asset = asset!("/assets/icon.png");
 pub const GITHUB_URL: &str = "https://github.com/vmux-ai/vmux";
@@ -53,72 +54,6 @@ pub fn Landing() -> Element {
             Hero {}
             Features {}
             Footer {}
-        }
-    }
-}
-
-#[component]
-fn Hero() -> Element {
-    use crate::hooks::{use_clipboard_copy, use_dmg_download, use_is_mac};
-    use dioxus_primitives::toast::{ToastOptions, use_toast};
-
-    let toast_api = use_toast();
-    let is_mac = use_is_mac();
-    let copy = use_clipboard_copy();
-    let download = use_dmg_download();
-
-    rsx! {
-        section { class: "text-center max-w-3xl mx-auto pt-16 pb-12 px-6 sm:pt-24 sm:pb-16 sm:px-8",
-            img {
-                src: ICON,
-                alt: "Vmux icon",
-                class: "w-32 h-32 mb-6 inline-block rounded-3xl",
-            }
-            h1 { class: "text-4xl sm:text-5xl font-bold mb-2 tracking-tight", "Vmux" }
-            p { class: "text-base sm:text-xl text-text-muted mb-10 max-w-xl mx-auto",
-                "Vibe Multiplexer — an agent-first workspace with a browser and IDE built in."
-            }
-            div { class: "flex flex-wrap justify-center gap-3 mb-6",
-                button {
-                    class: "inline-flex items-center px-6 py-3 rounded-lg text-base font-semibold border border-transparent bg-accent text-black cursor-pointer transition-colors hover:bg-accent-hover",
-                    onclick: move |_| {
-                        if is_mac {
-                            download(());
-                        } else {
-                            toast_api
-                                .info(
-                                    "Not supported".to_string(),
-                                    ToastOptions::new()
-                                        .description("Windows/Linux not supported yet — see GitHub Releases"),
-                                );
-                        }
-                    },
-                    "Download .dmg"
-                }
-                a {
-                    class: "inline-flex items-center px-6 py-3 rounded-lg text-base font-semibold no-underline border border-border bg-transparent text-text transition-colors hover:border-accent hover:text-accent",
-                    href: GITHUB_URL,
-                    target: "_blank",
-                    "GitHub"
-                }
-                Link {
-                    class: "inline-flex items-center px-6 py-3 rounded-lg text-base font-semibold no-underline border border-border bg-transparent text-text transition-colors hover:border-accent hover:text-accent",
-                    to: crate::Route::DocsIndex {},
-                    "Docs"
-                }
-            }
-            div { class: "inline-flex flex-col sm:flex-row items-center gap-2 sm:gap-3 bg-code-bg border border-border rounded-lg px-4 py-3 text-sm sm:text-base mb-4",
-                code { class: "font-mono text-accent", "{INSTALL_CMD}" }
-                button {
-                    class: "bg-accent text-black border-0 rounded px-3 py-1.5 text-sm font-semibold cursor-pointer transition-colors hover:bg-accent-hover",
-                    onclick: move |_| {
-                        copy(INSTALL_CMD.to_string());
-                        toast_api.success("Copied!".to_string(), ToastOptions::new());
-                    },
-                    "Copy"
-                }
-            }
-            p { class: "text-sm text-text-muted", "Requires macOS 13.0 (Ventura) or later." }
         }
     }
 }
